feat: Added stripe integration with example
Added all the integration with Stripe included in the payment app. This include an example in a button localhost:8000/checkout and the custom code to create an integration with the products "create_custom_checkout_session"

Mandatory to include the payment in the mai url, and the stripe keys in the settings.